pyproject.toml
[project]
name = "content-store"
version = "0.1.0"
description = "Simple marketplace on FastAPI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.122.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "python-jose>=3.5.0",
    "ruff>=0.14.6",
    "sqlalchemy>=2.0.44",
    "aiosqlite>=0.20.0",
    "pytest-asyncio>=1.3.0",
    "retry>=0.9.2",
]
